+++ rules-tests/SwiftMailer/Rector/MethodCall/SwiftCreateMessageToNewEmailRector/Fixture/skip_non_swift_mailer_create_message.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Symfony\Tests\Symfony53\Rector\MethodCall\SwiftCreateMessageToNewEmailRector\Fixture;
+namespace Rector\Symfony\Tests\SwiftMailer\Rector\MethodCall\SwiftCreateMessageToNewEmailRector\Fixture;
 
 final class SkipNonSwiftMailerCreateMessage
 {
