+++ rules-tests/SwiftMailer/Rector/MethodCall/SwiftSetBodyToHtmlPlainMethodCallRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Symfony\Symfony53\Rector\MethodCall\SwiftSetBodyToHtmlPlainMethodCallRector;
+use Rector\Symfony\SwiftMailer\Rector\MethodCall\SwiftSetBodyToHtmlPlainMethodCallRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(SwiftSetBodyToHtmlPlainMethodCallRector::class);
